+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # igraph C library changelog
 
+## [master]
+
 ## [0.10.9] - 2024-02-02
 
 ### Added