+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
  - `igraph_is_forest()` determined that a graph is not a directed forest, and the `roots` output parameter was set to `NULL`, it would incorrectly cache that the graph is also not an undirected forest.
+ - `igraph_spanner()` now correctly ignores edge directions, and no longer crashes on directed graphs.
 
 ### Other